Fixed Spring Data REST dependency to dedicated SNAPSHOT.
Current development of the 1.1.x branch of Spring Data REST introduces a few breaking changes and currently does not work as a drop in replacement. Thus, I am fixing the dependency version to the last snapshot know to work and maintain changes necessary to work with current 1.1.x snapshits in the branch sd-rest-next. That branch will be dropped once we fully upgraded to the latest Spring Data REST.