Bootify project.
Removed a lot of dependency configuration by moving to boot starter POMs. Removed entire JPA infrastructure configuration. Centralized application configuration and bootstrapping in Restbucks class. Removed the usage of a ApplicationContext hierarchy in favor of a single ApplicationContext. Customized the JPA mapping of Payment.order as order is a reserved keyword and fails the Hibernate version that we updated to.

Updated readme.md to reflect latest changes.