Switched to Moneta (JSR-354 reference implementation).
Removed custom MonetaryAmount value object in favor of the Moneta reference implementation of JSR-354. Extracted JacksonCustomizations into a separate configuration class to both register custom (de)serializers for the Money type as well as Jackson customizations needed in our domain model.

Changed domain types to be as immutable as possible.