#59 - Removed references to @Autowired after upgrade to Spring 4.3 RC1.
As of Spring 4.3 RC1 application components with a unique constructor is automatically autowired and the need for an explicit @Autowired annotation is removed. Thus, we can now stop using the onConstructor clause of @RequiredArgsConstructor in Lombok.

Changed test case for the order resource slightly due to the Content-Type header now containing the encoding as well. Removed ApplicationIntegrationTests as it breaks the context caching as ApplicatinoContexts are started and shut down so that the in-memory database created by tests using the test context framework is being shut down as well.