#93 - Improve design of payment triggering.
PaymentController now takes a complete JSON document to trigger the payment instead of a plain JSON String with a credit card number. This will allow using affordances APIs in a next step to produce the proper forms metadata.

Tweaked the serialization customizations so that we can add end up registering the regular expression used to verify the credit card number to be used as form metadata as well.