Revamp order placing process.
Prior to this commit, creation of orders was handled by Spring Data REST which, especially in the line items to be included in the order caused a lot of information provided that doesn't really make sense for a client to submit, such as the price of the item to be ordered. This has been superseded by moving product information into a Drink aggregate living in a "drinks" package. It is exposed by Spring Data REST primarily for reading purposes.

The order creation is now customized by an explicit OrderController.placeOrder(…) that takes a LocationAndDrinks close to what's to be described in a real coffee ordering experience. Drinks are submitted by pointing to Drink resources (described above). We make use of the AggregateReferenceResolvingModule newly introduced in Spring Data REST to automatically let Jackson map the URIs to Drink instances.

To advertise that step we make use of the – also newly introduced – support for HAL FORMS in Spring Data REST application. The form is crated in RootResourceModelProcessor attaching an affordance pointing to the custom controller method. Also, the affordance is enriched in JacksonCustomizations.halFormsConfiguration() with inline options for the location as well as remote options for the drinks array. The target for the remote option is DrinksOptions producing a filterable response in the expected format (list of prompt / value pairs).

A couple of renames across the codebase: repositories are now just pluralized variants of the aggregates they manage. PaymentException is just PaymentFailed.

The registration of the Jackson mixins needed has been moved onto the individual web packages by introducing the Mixins configuration type and  the RestbucksModule configured in JacksonCustomizations.