Architectural cleanups in drinks and payments.
Previously, CreditCardPayment used a many-to-one relationship to CreditCard which violates the relationship rules for aggregates. Added the jmolecules-archunit dependency and tweak DocumentationTest to actually verify the ApplicationModules instances to reveal that problem. It is now ultimately solved by rather using a jMolecules Association to model the relationship.

The verification of the ApplicationModules also revealed that Drinks and Order form a cyclic relationship as Drinks referred to the Milk and Size types residing in Order. This is solved by moving those into the Drinks module.